Add swager docs

+++ client/src/components/UserItem.js
@@ -2,7 +2,6 @@ import React, {useContext} from 'react';
 import {Card, Col} from "react-bootstrap";
 import Image from "react-bootstrap/Image";
 import {useNavigate} from "react-router-dom"
-import {USER_ROUTE} from "../utils/consts";
 import {observer} from "mobx-react-lite";
 import {FaEdit, FaTrash} from "react-icons/fa";
 import {deleteCompany} from "../http/companyAPI";
@@ -10,31 +9,33 @@ import {Context} from "../index";
 import IncomeChart from "./IncomeChart";
 import {deleteUser} from "../http/userAPI";
 
-const UserItem = observer(({users, onEdit}) => {
+const UserItem = observer(({userData, onEdit}) => {
     const navigate = useNavigate()
     const {user} = useContext(Context)
     const handleDelete = (id) => {
+        if(id===user.getCurrentUser.id){
+            alert("Нельзя удалить текущего пользователя")
+            return
+        }
         deleteUser(id).then(e => {
             alert('Пользователь удален')
-            user.setFlagRedraw(2)
+            user.setFlagRedrawUser(2)
         }).catch(e=>alert('Ошибка удаления'))
 
     }
-    const handleEdit = (id) => {
-        onEdit(id)
+    const handleEdit = (userId) => {
+        onEdit(userId)
     }
-
     return (
         <>
-            <Col md={3} className={"w-30 mt-3 d-flex align-items-stretch justify-items-between border-black"}
-                onClick={() => navigate(USER_ROUTE + '/' + users.id)}>
+            <Col md={3} className={"w-30 mt-3 d-flex align-items-stretch justify-items-between border-black"}>
             <Card style={{width: "300px", height: "350px", cursor: 'pointer'}} border={"black"}
                   className="p-2 d-flex justify-content-between">
-                <Image width="100%" height="50%" src={process.env.REACT_APP_API_URL + users.user_info.img}/>
+                <Image width="100%" height="50%" src={process.env.REACT_APP_API_URL + userData.user_info.img}/>
                 <div className="text-black-100 mt-1 d-flex flex-column   overflow-x-hidden">
-                    <div style={{fontSize: 20, fontWeight: "bold"}}>{users.name}...</div>
-                    <div className="d-flex flex-row"><p>Email: {users.email}$</p></div>
-                    <div className="d-flex flex-row"><p>Role: {users.role}$</p></div>
+                    <div style={{fontSize: 20, fontWeight: "bold"}}>{userData.name}...</div>
+                    <div className="d-flex flex-row"><p>Email: {userData.email}$</p></div>
+                    <div className="d-flex flex-row"><p>Role: {userData.role}$</p></div>
                 </div>
                 <div className="d-flex justify-content-between mt-2">
                     <FaEdit
@@ -42,7 +43,7 @@ const UserItem = observer(({users, onEdit}) => {
                         style={{cursor: 'pointer', marginRight: '10px'}}
                         onClick={(e) => {
                             e.stopPropagation();
-                            handleEdit(users.id);
+                            handleEdit(userData.id);
                         }}
                     />
                     <FaTrash
@@ -50,7 +51,7 @@ const UserItem = observer(({users, onEdit}) => {
                         style={{cursor: 'pointer'}}
                         onClick={(e) => {
                             e.stopPropagation();
-                            handleDelete(users.id);
+                            handleDelete(userData.id);
                         }}
                     />
                 </div>
